# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PIGE360 Self — Secretaria",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/v1/setup": {
      "post": {
        "tags": [
          "Autenticação e instalação"
        ],
        "summary": "Setup",
        "operationId": "setup_api_v1_setup_post",
        "parameters": [
          {
            "name": "x-setup-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Setup-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Setup"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Autenticação e instalação"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/schools": {
      "get": {
        "tags": [
          "Institucional e estrutura acadêmica"
        ],
        "summary": "Schools",
        "operationId": "schools_api_v1_schools_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Institucional e estrutura acadêmica"
        ],
        "summary": "Add School",
        "operationId": "add_school_api_v1_schools_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchoolInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Login": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "Login"
      },
...
      "SchoolInput": {
        "properties": {
          "company_id": {
            "type": "string",
            "title": "Company Id"
          },
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Name"
          },
          "address": {
            "type": "string",
            "maxLength": 400,
            "title": "Address",
            "default": ""
          },
          "phone": {
            "type": "string",
            "maxLength": 32,
            "title": "Phone",
            "default": ""
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "title": "Email",
            "default": ""
          },
          "document_policy": {
            "type": "string",
            "enum": [
              "warn",
              "block"
            ],
            "title": "Document Policy",
            "default": "warn"
          },
          "active": {
            "type": "boolean",
            "title": "Active",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "company_id",
          "name"
        ],
        "title": "SchoolInput"
      },
...
      "Setup": {
        "properties": {
          "admin_name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Admin Name"
          },
          "admin_email": {
            "type": "string",
            "format": "email",
            "title": "Admin Email"
          },
          "admin_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "title": "Admin Password"
          },
          "company_name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Company Name"
          },
          "company_document": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 24
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Document"
          },
          "school_name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "School Name"
          },
          "unit_name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Unit Name",
            "default": "Unidade principal"
          },
          "academic_year": {
            "type": "integer",
            "maximum": 2200.0,
            "minimum": 2000.0,
            "title": "Academic Year"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "admin_name",
          "admin_email",
          "admin_password",
          "company_name",
          "school_name",
          "academic_year"
        ],
        "title": "Setup"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}